requirements.txt
tekton==2.2.1
xmltodict==0.9.0
Jinja2==2.7.3
GoogleAppEngineCloudStorageClient==1.9.22.1